feat: add filter component with drag functionality and publication display
- Implemented a new FilterComponent for filtering careers with horizontal scrolling.
- Added TimeUtils class to calculate time elapsed since publication.
- Created PublicationWithFilesComponent and PublicationWithoutFilesComponent to display publications with and without files respectively.
- Enhanced UI with new styles and scrollbar hiding for better user experience.
- Added unit tests for all new components to ensure functionality.

+++ src/app/core/services/oauth/token.service.ts
@@ -1,26 +1,32 @@
 
-import { Injectable } from '@angular/core';
+import { isPlatformBrowser } from '@angular/common';
+import { Injectable, Inject, PLATFORM_ID } from '@angular/core';
 
 const TOKEN_KEY = 'AuthToken';
 @Injectable({
   providedIn: 'root'
 })
 export class TokenService  {
-  constructor() { }
+  constructor(@Inject(PLATFORM_ID) private platformId: Object) {}
 
   public getToken(): string {
-    if (typeof window !== 'undefined' && typeof sessionStorage !== 'undefined') {
+    if (isPlatformBrowser(this.platformId)) {
       return localStorage.getItem(TOKEN_KEY) || '';
     }
     return '';
   }
 
   public setToken(token: string): void {
-    localStorage.removeItem(TOKEN_KEY);
-    localStorage.setItem(TOKEN_KEY, token);
+    if (isPlatformBrowser(this.platformId)) {
+      localStorage.removeItem(TOKEN_KEY);
+      localStorage.setItem(TOKEN_KEY, token);
+    }
   }
-  logOut(): void {
-    localStorage.removeItem(TOKEN_KEY);
-    sessionStorage.clear();
+
+  public logOut(): void {
+    if (isPlatformBrowser(this.platformId)) {
+      localStorage.removeItem(TOKEN_KEY);
+      sessionStorage.clear();
+    }
   }
 }
